+++ crates/ink/tests/ui/trait_def/fail/message_immutable_payable.rs
@@ -1,0 +1,7 @@
+#[ink::trait_definition]
+pub trait TraitDefinition {
+    #[ink(message, payable)]
+    fn message(&self);
+}
+
+fn main() {}