+++ crates/ink/tests/ui/trait_def/pass/payable_message.rs
@@ -1,9 +1,6 @@
 #[ink::trait_definition]
 pub trait PayableDefinition {
     #[ink(message, payable)]
-    fn payable(&self);
-
-    #[ink(message, payable)]
     fn payable_mut(&mut self);
 
     #[ink(message)]
@@ -15,7 +12,6 @@ pub trait PayableDefinition {
 
 use ink::selector_id;
 
-const PAYABLE_ID: u32 = selector_id!("payable");
 const PAYABLE_MUT_ID: u32 = selector_id!("payable_mut");
 const UNPAYABLE_ID: u32 = selector_id!("unpayable");
 const UNPAYABLE_MUT_ID: u32 = selector_id!("unpayable_mut");
@@ -26,11 +22,7 @@ fn main() {
         TraitMessageInfo,
     };
     use ink_env::DefaultEnvironment;
-    assert!(
-        <<TraitDefinitionRegistry<DefaultEnvironment>
-            as PayableDefinition>::__ink_TraitInfo
-            as TraitMessageInfo<PAYABLE_ID>>::PAYABLE,
-    );
+
     assert!(
         <<TraitDefinitionRegistry<DefaultEnvironment>
             as PayableDefinition>::__ink_TraitInfo